logger: allow logger to be overridden with implementation accepting levels
The current logging implementation allows the logger to be overridden, but only
after filtering by log level, and changing the global log level at runtime is
not concurrency safe or convenient.

This patches the implementation so that SetLogger() can optionally provide an
implementation that supports LogAtLevel(), which will be called instead when
implemented by the inner logger.

This facilitates hooking up the Aerospike logger to our own logger while
preserving log levels.

Addresses ENG-3358